+++ Habit Management/Core/Component/RoundedCorner.swift
@@ -1,0 +1,21 @@
+//
+//  RoundedCorner.swift
+//  Habit Management
+//
+//  Created by 남경민 on 5/28/25.
+//
+
+import Foundation
+import SwiftUI
+
+struct RoundedCorner: Shape {
+    
+    var radius: CGFloat = .infinity
+    var corners: UIRectCorner = .allCorners
+    
+    func path(in rect: CGRect) -> Path {
+        let path = UIBezierPath(roundedRect: rect, byRoundingCorners: corners, cornerRadii: CGSize(width: radius, height: radius))
+        return Path(path.cgPath)
+    }
+    
+}
