+++ Habit Management/Core/Extension/View.swift
@@ -37,4 +37,5 @@ extension View {
     func cornerRadius(_ radius: CGFloat, corners: UIRectCorner) -> some View {
         clipShape( RoundedCorner(radius: radius, corners: corners) )
     }
+    
 }
